requirements.txt
docker-events==0.0.10
docker-py==1.4.0
python-ptrace==0.8.1
msgpack-python==0.4.6